server/src/main/java/ru/practicum/shareit/user/UserMapper.java
package ru.practicum.shareit.user;

import lombok.experimental.UtilityClass;
import ru.practicum.shareit.item.dto.ItemDto;
import ru.practicum.shareit.user.dto.UserDto;

@UtilityClass
public class UserMapper {

    public static UserDto userToUserDto(User user) {
        return new UserDto(
                user.getId(),
                user.getName(),
                user.getEmail()
        );
    }

    public static User userDtotoUser(UserDto userDto) {
        return new User(
                userDto.getId(),
                userDto.getName(),
                userDto.getEmail()
        );
    }

    public static ItemDto.Owner userDtoToOwner(UserDto userDto) {
        return new ItemDto.Owner(
                userDto.getId(),
                userDto.getName(),
                userDto.getEmail());
    }
}
